Move KeyboardInput file into keyboardinput folder
